Twilio multiple message distribution.py
PURPOSE OF THE PROJECT:

A script which allows the user to specify a number between a set of parameters which will determine the volume of messages they will receive. 

Testing the reliability of the Twilio API and it's capability of multiple message distributions.

Data within this project is secure and low threat to interception due to its use of the Operating System for storing account details locally to my computer environment. This way, as I have made my code open source, account details aren't at risk of theft/ manipulation.

LIBRARIES USED:
- Twilio.rest (API)
- Time
- OS