Add files via upload
Simple text file that is opened and read in read mode for the snake game to find the high score that has been achieved on the game.
This is a very simple demonstration of a high score log with the sole purpose to store only one value "high score". 
In this Snake game, the file reads the integer value stored for high score and overrides the value each time a new one is achieved.
For future development, instead of adding just a simple 1 line log of the highest score achieved, the user will be able to enter a their username to claim their victory on the game where the backend will keep track of the date & time the high score was achieved by the user and log it as a new entry for each record achieved. This log of high scores can be viewed by the users before loading a snake game however for utility purposes, I have kept this version as simple as possible, displaying my proficiency in OOP.