main.py
This file brings together all components of a clone I created of the famous arcade game "Pong". 
This file imports all external classes into one file to create object instances onto a GUI to display to the user who can interact with the 2 paddles with another player.

I have decided to keep this project as simple as possible to display simple functionality however there are some improvements to be added that would make the game more efficient.

- A choice for 1 player, 2 players or "online mode". 

This version has the 2 paddles controlled by a single keyboard, where 2 players can play each other when using the same keyboard.

1 player mode: this can be used for players who want to play against a computer. An AI can be programmed and written in an extra file to make moves like a player would in order to replicate the feel of a 2 player game.

Online mode: this can be used for players who want to play against other players. This would be achieved through implementing a server using "threads & sockets" with a client and server to host the game on where 2 players will see the game movement simultaneously whilst not being on the same system. 

These changes will be added at a later date.